Moving TeamOverviewListItem
